+++ packages/block-library/src/image/edit.native.js
@@ -12,29 +12,44 @@ import {
 	requestImageFailedRetryDialog,
 	requestImageUploadCancelDialog,
 } from 'react-native-gutenberg-bridge';
+
+/**
+ * WordPress dependencies
+ */
 import {
-	map,
-	compact,
-} from 'lodash';
+	Toolbar,
+	ToolbarButton,
+	Spinner,
+	Dashicon,
+} from '@wordpress/components';
+import {
+	MediaPlaceholder,
+	RichText,
+	BlockControls,
+	InspectorControls,
+	BottomSheet,
+	Picker,
+} from '@wordpress/editor';
+import { __ } from '@wordpress/i18n';
+import { isURL } from '@wordpress/url';
 
 /**
  * Internal dependencies
  */
-import { MediaPlaceholder, RichText, BlockControls, InspectorControls, BottomSheet } from '@wordpress/editor';
-import { Toolbar, ToolbarButton, Spinner, Dashicon } from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
 import ImageSize from './image-size';
-import { isURL } from '@wordpress/url';
 import styles from './styles.scss';
-import { compose } from '@wordpress/compose';
-import { withSelect } from '@wordpress/data';
 
 const MEDIA_UPLOAD_STATE_UPLOADING = 1;
 const MEDIA_UPLOAD_STATE_SUCCEEDED = 2;
 const MEDIA_UPLOAD_STATE_FAILED = 3;
 const MEDIA_UPLOAD_STATE_RESET = 4;
 
+const MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_CHOOSE_FROM_DEVICE = 'choose_from_device';
+const MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_TAKE_PHOTO = 'take_photo';
+const MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_WORD_PRESS_LIBRARY = 'wordpress_media_library';
+
 const LINK_DESTINATION_CUSTOM = 'custom';
+const LINK_DESTINATION_NONE = 'none';
 
 class ImageEdit extends React.Component {
 	constructor( props ) {
@@ -56,6 +71,7 @@ class ImageEdit extends React.Component {
 		this.updateImageURL = this.updateImageURL.bind( this );
 		this.onSetLinkDestination = this.onSetLinkDestination.bind( this );
 		this.onImagePressed = this.onImagePressed.bind( this );
+		this.onClearSettings = this.onClearSettings.bind( this );
 	}
 
 	componentDidMount() {
@@ -154,14 +170,20 @@ class ImageEdit extends React.Component {
 		} );
 	}
 
-	getImageSizeOptions() {
-		const { imageSizes } = this.props;
-		return compact( map( imageSizes, ( { label, slug } ) => {
-			return {
-				value: this.props.attributes.url + slug, //temporary url
-				label,
-			};
-		} ) );
+	onClearSettings() {
+		this.props.setAttributes( {
+			alt: '',
+			linkDestination: LINK_DESTINATION_NONE,
+			href: undefined,
+		} );
+	}
+
+	getMediaOptionsItems() {
+		return [
+			{ icon: 'wordpress-alt', value: MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_CHOOSE_FROM_DEVICE, label: __( 'Choose from device' ) },
+			{ icon: 'camera', value: MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_TAKE_PHOTO, label: __( 'Take a Photo' ) },
+			{ icon: 'format-image', value: MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_WORD_PRESS_LIBRARY, label: __( 'WordPress Media Library' ) },
+		];
 	}
 
 	render() {
@@ -176,33 +198,23 @@ class ImageEdit extends React.Component {
 			} );
 		};
 
-		if ( ! url ) {
-			const onMediaUploadButtonPressed = () => {
-				requestMediaPickFromDeviceLibrary( ( mediaId, mediaUri ) => {
-					if ( mediaUri ) {
-						this.addMediaUploadListener( );
-						setAttributes( { url: mediaUri, id: mediaId } );
-					}
-				} );
-			};
+		const onMediaUploadButtonPressed = () => {
+			requestMediaPickFromDeviceLibrary( ( mediaId, mediaUri ) => {
+				if ( mediaUri ) {
+					this.addMediaUploadListener( );
+					setAttributes( { url: mediaUri, id: mediaId } );
+				}
+			} );
+		};
 
-			const onMediaCaptureButtonPressed = () => {
-				requestMediaPickFromDeviceCamera( ( mediaId, mediaUri ) => {
-					if ( mediaUri ) {
-						this.addMediaUploadListener( );
-						setAttributes( { url: mediaUri, id: mediaId } );
-					}
-				} );
-			};
-
-			return (
-				<MediaPlaceholder
-					onUploadMediaPressed={ onMediaUploadButtonPressed }
-					onMediaLibraryPressed={ onMediaLibraryButtonPressed }
-					onCapturePhotoPressed={ onMediaCaptureButtonPressed }
-				/>
-			);
-		}
+		const onMediaCaptureButtonPressed = () => {
+			requestMediaPickFromDeviceCamera( ( mediaId, mediaUri ) => {
+				if ( mediaUri ) {
+					this.addMediaUploadListener( );
+					setAttributes( { url: mediaUri, id: mediaId } );
+				}
+			} );
+		};
 
 		const onImageSettingsButtonPressed = () => {
 			this.setState( { showSettings: true } );
@@ -212,17 +224,21 @@ class ImageEdit extends React.Component {
 			this.setState( { showSettings: false } );
 		};
 
+		let picker;
+
+		const onMediaOptionsButtonPressed = () => {
+			picker.presentPicker();
+		};
+
 		const toolbarEditButton = (
 			<Toolbar>
 				<ToolbarButton
 					label={ __( 'Edit image' ) }
 					icon="edit"
-					onClick={ onMediaLibraryButtonPressed }
+					onClick={ onMediaOptionsButtonPressed }
 				/>
 			</Toolbar>
 		);
-
-		const sizeOptions = this.getImageSizeOptions();
 
 		const getInspectorControls = () => (
 			<BottomSheet
@@ -239,13 +255,6 @@ class ImageEdit extends React.Component {
 					autoCapitalize="none"
 					autoCorrect={ false }
 				/>
-				<BottomSheet.PickerCell
-					icon="editor-expand"
-					label={ __( 'Image Size' ) }
-					value={ 'Large' } // Temporary for UI implementation.
-					options={ sizeOptions }
-					onChangeValue={ () => {} } // Temporary for UI implementation.
-				/>
 				<BottomSheet.Cell
 					icon={ 'editor-textcolor' }
 					label={ __( 'Alt Text' ) }
@@ -254,13 +263,43 @@ class ImageEdit extends React.Component {
 					onChangeValue={ this.updateAlt }
 				/>
 				<BottomSheet.Cell
-					label={ __( 'Reset to Original' ) }
-					labelStyle={ styles.resetSettingsButton }
+					label={ __( 'Clear All Settings' ) }
+					labelStyle={ styles.clearSettingsButton }
 					drawSeparator={ false }
-					onPress={ () => {} }
+					onPress={ this.onClearSettings }
 				/>
 			</BottomSheet>
 		);
+
+		const mediaOptions = this.getMediaOptionsItems();
+
+		const getMediaOptions = () => (
+			<Picker
+				hideCancelButton={ true }
+				ref={ ( instance ) => picker = instance }
+				options={ mediaOptions }
+				onChange={ ( value ) => {
+					if ( value === MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_CHOOSE_FROM_DEVICE ) {
+						onMediaUploadButtonPressed();
+					} else if ( value === MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_TAKE_PHOTO ) {
+						onMediaCaptureButtonPressed();
+					} else if ( value === MEDIA_UPLOAD_BOTTOM_SHEET_VALUE_WORD_PRESS_LIBRARY ) {
+						onMediaLibraryButtonPressed();
+					}
+				} }
+			/>
+		);
+
+		if ( ! url ) {
+			return (
+				<View style={ { flex: 1 } } >
+					{ getMediaOptions() }
+					<MediaPlaceholder
+						onMediaOptionsPressed={ onMediaOptionsButtonPressed }
+					/>
+				</View>
+			);
+		}
 
 		const showSpinner = this.state.isUploadInProgress;
 		const opacity = this.state.isUploadInProgress ? 0.3 : 1;
@@ -308,6 +347,7 @@ class ImageEdit extends React.Component {
 							return (
 								<View style={ { flex: 1 } } >
 									{ getInspectorControls() }
+									{ getMediaOptions() }
 									<ImageBackground
 										style={ { width: finalWidth, height: finalHeight, opacity } }
 										resizeMethod="scale"
@@ -343,15 +383,4 @@ class ImageEdit extends React.Component {
 	}
 }
 
-export default compose( [
-	withSelect( ( select ) => {
-		const { getEditorSettings } = select( 'core/editor' );
-		const { maxWidth, isRTL, imageSizes } = getEditorSettings();
-
-		return {
-			maxWidth,
-			isRTL,
-			imageSizes,
-		};
-	} ),
-] )( ImageEdit );
+export default ImageEdit;