+++ packages/block-library/src/table/index.js
@@ -95,7 +95,7 @@ export const settings = {
 	},
 
 	styles: [
-		{ name: 'regular', label: _x( 'Regular', 'block style' ), isDefault: true },
+		{ name: 'regular', label: _x( 'Default', 'block style' ), isDefault: true },
 		{ name: 'stripes', label: __( 'Stripes' ) },
 	],