+++ packages/e2e-test-utils/src/click-on-more-menu-item.js
@@ -4,11 +4,6 @@
 import { first } from 'lodash';
 
 /**
- * Internal dependencies
- */
-import { waitForAnimation } from './wait-for-animation';
-
-/**
  * Clicks on More Menu item, searches for the button with the text provided and clicks it.
  *
  * @param {string} buttonLabel The label to search the button for.
@@ -17,7 +12,6 @@ export async function clickOnMoreMenuItem( buttonLabel ) {
 	await expect( page ).toClick(
 		'.edit-post-more-menu [aria-label="Show more tools & options"]'
 	);
-	await waitForAnimation();
 	const moreMenuContainerSelector =
 		'//*[contains(concat(" ", @class, " "), " edit-post-more-menu__content ")]';
 	let elementToClick = first( await page.$x(