+++ packages/e2e-tests/specs/links.test.js
@@ -6,9 +6,7 @@ import {
 	getEditedPostContent,
 	createNewPost,
 	pressKeyWithModifier,
-	pressKeyTimes,
 	insertBlock,
-	waitForAnimation,
 } from '@wordpress/e2e-test-utils';
 
 /**
@@ -41,7 +39,6 @@ describe( 'Links', () => {
 
 		// Click on the Link button
 		await page.click( 'button[aria-label="Link"]' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
@@ -66,7 +63,6 @@ describe( 'Links', () => {
 
 		// Press Cmd+K to insert a link
 		await pressKeyWithModifier( 'primary', 'K' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
@@ -91,7 +87,6 @@ describe( 'Links', () => {
 
 		// Press Cmd+K to insert a link
 		await pressKeyWithModifier( 'primary', 'K' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
@@ -134,7 +129,6 @@ describe( 'Links', () => {
 
 		// Click on the Link button
 		await page.click( 'button[aria-label="Link"]' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
@@ -156,7 +150,6 @@ describe( 'Links', () => {
 
 		// Click on the Link button
 		await page.click( 'button[aria-label="Link"]' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
@@ -213,7 +206,6 @@ describe( 'Links', () => {
 		await clickBlockAppender();
 		await page.keyboard.type( 'Text' );
 		await page.click( 'button[aria-label="Link"]' );
-		await waitForAnimation();
 
 		// Typing "left" should not close the dialog
 		await page.keyboard.press( 'ArrowLeft' );
@@ -236,7 +228,6 @@ describe( 'Links', () => {
 		await moveMouse();
 		await page.waitForSelector( 'button[aria-label="Link"]' );
 		await page.click( 'button[aria-label="Link"]' );
-		await waitForAnimation();
 
 		// Typing "left" should not close the dialog
 		await page.keyboard.press( 'ArrowLeft' );
@@ -252,7 +243,8 @@ describe( 'Links', () => {
 	it( 'can be edited with collapsed selection', async () => {
 		await createAndReselectLink();
 		// Make a collapsed selection inside the link
-		await pressKeyTimes( 'ArrowRight', 3 );
+		await page.keyboard.press( 'ArrowLeft' );
+		await page.keyboard.press( 'ArrowRight' );
 		await moveMouse();
 		await page.click( 'button[aria-label="Edit"]' );
 		await waitForAutoFocus();
@@ -293,14 +285,12 @@ describe( 'Links', () => {
 		await page.keyboard.type( 'This is Gutenberg' );
 		await pressKeyWithModifier( 'shiftAlt', 'ArrowLeft' );
 		await page.click( 'button[aria-label="Link"]' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
 
 		await page.keyboard.type( titleText );
 		await page.waitForSelector( '.editor-url-input__suggestion' );
-		await waitForAnimation();
 		const autocompleteSuggestions = await page.$x( `//*[contains(@class, "editor-url-input__suggestion")]//button[contains(text(), '${ titleText }')]` );
 
 		// Expect there to be some autocomplete suggestions.
@@ -337,14 +327,12 @@ describe( 'Links', () => {
 
 		// Press Cmd+K to insert a link
 		await pressKeyWithModifier( 'primary', 'K' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
 
 		await page.keyboard.type( titleText );
 		await page.waitForSelector( '.editor-url-input__suggestion' );
-		await waitForAnimation();
 		const autocompleteSuggestions = await page.$x( `//*[contains(@class, "editor-url-input__suggestion")]//button[contains(text(), '${ titleText }')]` );
 
 		// Expect there to be some autocomplete suggestions.
@@ -375,7 +363,6 @@ describe( 'Links', () => {
 
 		// Press Cmd+K to insert a link
 		await pressKeyWithModifier( 'primary', 'K' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
@@ -384,7 +371,6 @@ describe( 'Links', () => {
 		// Trigger the autocomplete suggestion list and select the first suggestion.
 		await page.keyboard.type( titleText );
 		await page.waitForSelector( '.editor-url-input__suggestion' );
-		await waitForAnimation();
 		await page.keyboard.press( 'ArrowDown' );
 
 		// Expect the the escape key to dismiss the popover when the autocomplete suggestion list is open.
@@ -393,7 +379,6 @@ describe( 'Links', () => {
 
 		// Press Cmd+K to insert a link
 		await pressKeyWithModifier( 'primary', 'K' );
-		await waitForAnimation();
 
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
@@ -427,7 +412,6 @@ describe( 'Links', () => {
 		await page.keyboard.type( 'This is Gutenberg' );
 		await pressKeyWithModifier( 'shiftAlt', 'ArrowLeft' );
 		await pressKeyWithModifier( 'primary', 'K' );
-		await waitForAnimation();
 		await waitForAutoFocus();
 		await page.keyboard.type( URL );
 		await page.keyboard.press( 'Enter' );
@@ -445,7 +429,6 @@ describe( 'Links', () => {
 		// Press Cmd+K to edit the link and the url-input should become
 		// focused with the value previously inserted.
 		await pressKeyWithModifier( 'primary', 'K' );
-		await waitForAnimation();
 		await waitForAutoFocus();
 		const activeElementParentClasses = await page.evaluate( () => Object.values( document.activeElement.parentElement.classList ) );
 		expect( activeElementParentClasses ).toContain( 'editor-url-input' );
@@ -458,7 +441,6 @@ describe( 'Links', () => {
 		await page.keyboard.type( 'This is Gutenberg' );
 		await pressKeyWithModifier( 'shiftAlt', 'ArrowLeft' );
 		await pressKeyWithModifier( 'primary', 'K' );
-		await waitForAnimation();
 		await waitForAutoFocus();
 		await page.keyboard.type( 'http://#test.com' );
 		await page.keyboard.press( 'Enter' );
@@ -482,8 +464,6 @@ describe( 'Links', () => {
 
 		// Click on the Link button
 		await page.click( 'button[aria-label="Link"]' );
-		await waitForAnimation();
-
 		// Wait for the URL field to auto-focus
 		await waitForAutoFocus();
 
@@ -510,7 +490,6 @@ describe( 'Links', () => {
 		// Select "WordPress".
 		await pressKeyWithModifier( 'shiftAlt', 'ArrowLeft' );
 		await pressKeyWithModifier( 'primary', 'k' );
-		await waitForAnimation();
 		await waitForAutoFocus();
 		await page.keyboard.type( 'w.org' );
 		// Navigate to the settings toggle.